+++ src/file/domain/file_meta.rs
@@ -2,7 +2,10 @@ use std::{fs::Metadata, path::Path};
 
 use serde::{Deserialize, Serialize};
 
-use crate::time::Timestamp;
+use crate::{
+    file::{is_image, is_movie},
+    time::Timestamp,
+};
 
 // get meta infor from fs::Metadata
 // because only one IO operation per Metadata fetch,
@@ -14,6 +17,9 @@ pub struct FileMeta {
     pub modified: u64, // Timestamp
     pub created: u64,  // Timestamp
     pub size: u64,
+
+    pub is_image: bool,
+    pub is_movie: bool,
 }
 
 impl Default for FileMeta {
@@ -25,17 +31,31 @@ impl Default for FileMeta {
             modified: 0,
             created: 0,
             size: 0,
+            is_image: false,
+            is_movie: false,
         }
     }
 }
 
-impl From<Metadata> for FileMeta {
-    fn from(meta: Metadata) -> Self {
+impl From<&Path> for FileMeta {
+    fn from(path: &Path) -> Self {
+        let meta = match path.metadata() {
+            Ok(meta) => meta,
+            Err(_) => return FileMeta::default(),
+        };
+
         let modified = meta
             .modified()
             .map(Timestamp::from_system_time)
             .unwrap_or(0);
         let created = meta.created().map(Timestamp::from_system_time).unwrap_or(0);
+
+        let (is_image_, is_movie_) = if let Some(ext) = path.extension() {
+            let ext = ext.to_string_lossy().to_lowercase();
+            (is_image(&ext), is_movie(&ext))
+        } else {
+            (false, false)
+        };
 
         FileMeta {
             is_dir: meta.is_dir(),
@@ -44,15 +64,8 @@ impl From<Metadata> for FileMeta {
             modified,
             created,
             size: meta.len(),
-        }
-    }
-}
-
-impl From<&Path> for FileMeta {
-    fn from(path_buf: &Path) -> Self {
-        match path_buf.metadata() {
-            Ok(meta) => FileMeta::from(meta),
-            Err(_) => FileMeta::default(),
+            is_image: is_image_,
+            is_movie: is_movie_,
         }
     }
 }